Unnecessary folder deleted

+++ MR_SYSTEM_Jinal/userapp/views.py
@@ -13,31 +13,13 @@ from sklearn.model_selection import train_test_split
 from sklearn.svm import LinearSVC
 
 # Create your views here.
-""" movies = {0:{'Movie':"Genius",'Review':"-",'Rating':0}, 1:{'Movie':"Mission Mangal", 'Review':"-",'Rating':0}, 
-2:{'Movie':"3 idiots",'Review':"-",'Rating':0}, 3:{'Movie':"URI",'Review':"-",'Rating':0}}
-tfidf= []
-clf= []
 
-def display_movies(request):
-    c={}
-    c.update(csrf(request))
-    df = pd.read_excel('../Templates/AmazonSDPDataset_try.ods', engine='odf', usecols= ['reviewText','overall'])
-    df['reviewText']= df['reviewText'].apply(lambda x: get_clean(x))
-
-    global tfidf
-    tfidf = TfidfVectorizer(analyzer='word') # append ngram_range=(1,5), analyzer='char'
-    X = df['reviewText']
-    y = df['overall']
-    X = tfidf.fit_transform(X)
-    X_train, X_test, y_train, y_test = train_test_split(X, y, test_size = 0.25, random_state = 0)
-
-    global clf
-    clf = LinearSVC(C = 0.1, class_weight= 'balanced') #append C = 10, class_weight= 'balanced'
-    clf.fit(X_train, y_train)
-    return render(request,'display_movies.html',{'c':c,'movies':movies})
+tfidf=[]
+clf =[]
+review=""
 
 def get_clean(x):
-    x = str(x).lower().replace('\\','').replace('_',' ').replace(',',' ')
+    x = str(x).lower().replace('\\', '').replace('_', ' ').replace(',','')
     x = ps.cont_exp(x)
     x = ps.remove_emails(x)
     x = ps.remove_urls(x)
@@ -47,21 +29,111 @@ def get_clean(x):
     x = re.sub("(.)\\1{2,}", "\\1", x)
     return x
 
-def calculate_rating(request):
-    review_text = request.POST.get('review','')
-    key =  int(request.POST.get('key',''))
-    movies[key]['Review']= review_text
+def readExcel(request):
 
-    global clf
+    df = pd.read_excel('~/SDP_Project/MR_SYSTEM/userapp/templates/AmazonSDPDataset_original.ods', engine='odf', usecols= ['reviewText','overall'])
+    df['reviewText'] = df['reviewText'].apply(lambda x: get_clean(x))
+
     global tfidf
-    review = tfidf.transform([get_clean(review_text)])
-    rating = clf.predict(review)
+    tfidf = TfidfVectorizer(analyzer='word')
+    X = tfidf.fit_transform(df['reviewText'])
+    Y = df['overall']
+    X_train, X_test, Y_train, Y_test = train_test_split(X,Y,test_size=0.25, random_state=0)
+    global clf
+    clf = LinearSVC(C=0.1, class_weight='balanced')
+    clf.fit(X_train,Y_train)
+    return HttpResponseRedirect('/user/user_home/')
 
-    movies[key]['Rating'] = rating[0]
-    c={}
-    c.update(csrf(request))
-    return render(request,'display_movies.html',{'c':c, 'movies':movies, 'rating':rating[0], 'review':review_text})
-"""
+
+
+def user_home(request):
+    if "user" in request.session:
+        cid = request.session["user"]
+        star = request.GET.get('star','')
+        filter = False
+        user = User.objects.filter(ID=cid)
+        if user.exists():
+            user = User.objects.get(ID=cid)
+            username = user.name
+            movies = Movie.objects.all()
+            if star != "":
+                movies = Movie.objects.filter(rating__range=(float(star)-1,float(star))).order_by('-rating')
+                filter = True
+            if movies.exists():
+                return render(request,'user_home.html',{'movielist':movies,'nomovie':False,'filter':filter,'username':username})
+            else:
+                return render(request,'user_home.html',{'movielist':movies,'nomovie':True,'filter':filter,'username':username})
+    return HttpResponseRedirect('/login/')
+
+def calculateRating(request):
+    if "user" in request.session:
+        id=request.POST.get('movieid','')
+        reviewText=request.POST.get('reviewText','')
+        global review
+        review = reviewText
+        reviewText = get_clean(reviewText)
+    
+        global clf
+        global tfidf
+        reviewText = tfidf.transform([reviewText])
+        rating = clf.predict(reviewText)
+        return HttpResponseRedirect('/user/addReview?rating='+str(rating[0])+'&movieid='+str(id))
+    return HttpResponseRedirect('/login/')
+
+def addReview(request):
+    if "user" in request.session:
+        rating = request.GET.get('rating','')
+        id = request.GET.get('movieid','')
+        global review
+        reviewText = review
+        Rating = rating
+        DateTime = datetime.now()
+        uid = request.session["user"]
+        user = User.objects.get(ID=uid)
+        movie = Movie.objects.get(ID=id)
+
+        new_review = Review(reviewText=reviewText,rating=Rating,dateTime=DateTime,mid=movie,uid=user)
+        new_review.save()
+
+        avg_rating(new_review.mid_id) #update avg movie rating
+
+        movie = Movie.objects.get(ID=id)
+        movie.releasedDate = (movie.releasedDate).strftime("%d-%m-%Y")
+        movie.duration = (movie.duration).strftime("%H:%M")
+
+        reviews = Review.objects.filter(mid=movie)
+        return HttpResponseRedirect('/user/showmovie?movieid='+str(id))
+    return HttpResponseRedirect('/login/')
+
+def showmovie(request):
+    if "user" in request.session:
+        cid = request.session["user"]
+        m =request.GET.get('m','')
+        if m == "":
+            m= False
+        sortby= request.GET.get('star','')
+        id = request.GET.get('movieid','')
+
+        movie = Movie.objects.get(ID=id)
+        user = User.objects.get(ID= cid)
+        added = False
+
+        myreviews= Review.objects.filter(mid_id= movie, uid_id=user) #don't allow user to add review if already added
+        if myreviews.exists():
+            added = True
+        movie.releasedDate = (movie.releasedDate).strftime("%Y-%m-%d")
+        movie.duration = (movie.duration).strftime("%H:%M")
+       
+        if sortby != '':
+            reviews= Review.objects.filter(mid=id, rating= sortby)
+        else:
+            reviews = Review.objects.filter(mid=id)
+        sortedReviews = sorted(
+            reviews,
+            key=lambda x: x.dateTime, reverse=True
+        )
+        return render(request,'showmovie.html',{'movie':movie,'reviews':sortedReviews,'currentuserid':cid, 'added':added,'mr':m})
+    return HttpResponseRedirect('/login/')
 
 def my_reviews(request):
     c= {}
@@ -69,60 +141,56 @@ def my_reviews(request):
     if 'user' in request.session:
         userid = request.session["user"]
         getuser = User.objects.get(ID = userid)
-        #console.log(user.email)
-        reviews = Review.objects.filter(uid_id= getuser)
+
+        reviews = Review.objects.filter(uid_id= getuser).order_by('-dateTime')
         if reviews.exists():
             return render(request,'myreviews.html',{'reviews':reviews,'c':c})
         else:
-            return render(request,'myreviews.html',{'reviews':reviews, 'msg':'No reviews available','c':c})
-
+            return render(request,'myreviews.html',{'reviews':reviews, 'msg':'You have not added any reviews yet..!!','c':c})
+    return HttpResponseRedirect('/login/')
 
 def update_review(request):
+    if 'user' not in request.session:
+        return HttpResponseRedirect('/login/')
     rid = request.POST.get('id','')
     if rid:
         review = Review.objects.get(ID= rid)
         review.reviewText = request.POST.get('new-rw','')
         review.dateTime = datetime.now()
-        store_rating(review)
+
+        global clf
+        global tfidf
+        rw = tfidf.transform([get_clean(review.reviewText)])
+        review.rating = clf.predict(rw)
+        review.save()
+        avg_rating(review.mid_id)
     return HttpResponseRedirect('/user/reviews/')
 
 def delete_review(request):
+    if 'user' not in request.session:
+        return HttpResponseRedirect('/login/')
+    
     rid = request.GET.get('id','')
     if rid:
         review= Review.objects.get(ID= rid)
         movie_id = review.mid_id
         review.delete()
-
-        movie =  Movie.objects.get(ID= movie_id)
-        reviews = Review.objects.filter(mid_id= movie) #update average rating of movie
-        rating = 0.0
-
-        if reviews.exists():
-            for r in reviews:
-                rating += r.rating
-            rating /= len(reviews)
-
-        movie.rating = rating
-        movie.save()
+        avg_rating(movie_id)
     return HttpResponseRedirect('/user/reviews/')
 
-def store_rating(review):
-    "function to calculate rating, average rating & to update it in database"
-    global clf
-    global tfidf
-    rw = tfidf.transform([get_clean(review.reviewText)])
-    review.rating = clf.predict(rw)
-    review.save()
+def avg_rating(movie_id):
+    "function to calculate average rating & to update it in database"
 
-    movie =  Movie.objects.get(ID= review.mid_id)
+    movie =  Movie.objects.get(ID= movie_id)
     reviews = Review.objects.filter(mid_id= movie) #update average rating of movie
     rating = 0.0
         
-    for r in reviews:
-        rating += r.rating
-    rating /= len(reviews)
+    if reviews.exists():
+        for r in reviews:
+            rating += r.rating
+        rating /= len(reviews)
 
-    movie.rating = rating
+    movie.rating = round(rating,1) #store rating in y.x format- 1 decimal point
     movie.save()
     return
